+++ src/error.rs
@@ -1,10 +1,10 @@
-// Copyright (C) 2019, 2020 O.S. Systems Sofware LTDA
+// Copyright (C) 2019, 2020 O.S. Systems Software LTDA
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use crate::ffi;
 use derive_more::{Display, Error, From};
-use std::ffi::CStr;
+use std::{ffi::CStr, io};
 
 pub type Result<T> = std::result::Result<T, Error>;
 
@@ -14,10 +14,10 @@ pub enum Error {
     #[display(fmt = "Extraction error: '{}'", _0)]
     Extraction(#[error(not(source))] String),
 
-    #[display(fmt = "Io error: '{}'", _0)]
-    Io(std::io::Error),
+    #[display(transparent)]
+    Io(io::Error),
 
-    #[display(fmt = "Utf error: '{}'", _0)]
+    #[display(transparent)]
     Utf(std::str::Utf8Error),
 
     #[cfg(feature = "tokio_support")]
@@ -26,12 +26,6 @@ pub enum Error {
 
     #[display(fmt = "Error to create the archive struct, is null")]
     NullArchive,
-
-    #[display(fmt = "The entry is null, failed to set the pathname")]
-    NullEntry,
-
-    #[display(fmt = "File not found")]
-    FileNotFound,
 
     #[display(fmt = "Unknown error")]
     Unknown,
@@ -50,14 +44,18 @@ impl From<*mut ffi::archive> for Error {
     fn from(input: *mut ffi::archive) -> Self {
         unsafe {
             let error_string = ffi::archive_error_string(input);
-            let errno = ffi::archive_errno(input);
             if !error_string.is_null() {
-                Error::Extraction(CStr::from_ptr(error_string).to_string_lossy().to_string())
-            } else if errno != 0 {
-                std::io::Error::from_raw_os_error(errno).into()
-            } else {
-                Error::Unknown
+                return Error::Extraction(
+                    CStr::from_ptr(error_string).to_string_lossy().to_string(),
+                );
+            }
+
+            let errno = ffi::archive_errno(input);
+            if errno != 0 {
+                return io::Error::from_raw_os_error(errno).into();
             }
         }
+
+        Error::Unknown
     }
 }